@@ -1,32 +1,28 @@
-#ifndef FLUTTER_WINDOW_H_
-#define FLUTTER_WINDOW_H_
+#ifndef RUNNER_FLUTTER_WINDOW_H_
+#define RUNNER_FLUTTER_WINDOW_H_
 
 #include <flutter/dart_project.h>
 #include <flutter/flutter_view_controller.h>
 
-#include "run_loop.h"
-#include "win32_window.h"
-
 #include <memory>
+
+#include "win32_window.h"
 
 // A window that does nothing but host a Flutter view.
 class FlutterWindow : public Win32Window {
  public:
-  // Creates a new FlutterWindow driven by the |run_loop|, hosting a
-  // Flutter view running |project|.
-  explicit FlutterWindow(RunLoop* run_loop,
-                         const flutter::DartProject& project);
+  // Creates a new FlutterWindow hosting a Flutter view running |project|.
+  explicit FlutterWindow(const flutter::DartProject& project);
   virtual ~FlutterWindow();
 
  protected:
   // Win32Window:
-  void OnCreate() override;
+  bool OnCreate() override;
   void OnDestroy() override;
+  LRESULT MessageHandler(HWND window, UINT const message, WPARAM const wparam,
+                         LPARAM const lparam) noexcept override;
 
  private:
-  // The run loop driving events for this window.
-  RunLoop* run_loop_;
-
   // The project to run.
   flutter::DartProject project_;
 
@@ -34,4 +30,4 @@ class FlutterWindow : public Win32Window {
   std::unique_ptr<flutter::FlutterViewController> flutter_controller_;
 };
 
-#endif  // FLUTTER_WINDOW_H_
+#endif  // RUNNER_FLUTTER_WINDOW_H_
