@@ -3,34 +3,40 @@
 #include <windows.h>
 
 #include "flutter_window.h"
-#include "run_loop.h"
-#include "window_configuration.h"
+#include "utils.h"
 
-int APIENTRY wWinMain(_In_ HINSTANCE instance,
-                      _In_opt_ HINSTANCE prev,
-                      _In_ wchar_t* command_line,
-                      _In_ int show_command) {
+int APIENTRY wWinMain(_In_ HINSTANCE instance, _In_opt_ HINSTANCE prev,
+                      _In_ wchar_t *command_line, _In_ int show_command) {
   // Attach to console when present (e.g., 'flutter run') or create a
   // new console when running with a debugger.
   if (!::AttachConsole(ATTACH_PARENT_PROCESS) && ::IsDebuggerPresent()) {
-    ::AllocConsole();
+    CreateAndAttachConsole();
   }
 
-  // Initialize COM, so that it is available for use in the library and/or plugins.
+  // Initialize COM, so that it is available for use in the library and/or
+  // plugins.
   ::CoInitializeEx(nullptr, COINIT_APARTMENTTHREADED);
 
-  RunLoop run_loop;
-
   flutter::DartProject project(L"data");
-  FlutterWindow window(&run_loop, project);
-  Win32Window::Point origin(kFlutterWindowOriginX, kFlutterWindowOriginY);
-  Win32Window::Size size(kFlutterWindowWidth, kFlutterWindowHeight);
-  if (!window.CreateAndShow(kFlutterWindowTitle, origin, size)) {
+
+  std::vector<std::string> command_line_arguments =
+      GetCommandLineArguments();
+
+  project.set_dart_entrypoint_arguments(std::move(command_line_arguments));
+
+  FlutterWindow window(project);
+  Win32Window::Point origin(10, 10);
+  Win32Window::Size size(1280, 720);
+  if (!window.CreateAndShow(L"example", origin, size)) {
     return EXIT_FAILURE;
   }
   window.SetQuitOnClose(true);
 
-  run_loop.Run();
+  ::MSG msg;
+  while (::GetMessage(&msg, nullptr, 0, 0)) {
+    ::TranslateMessage(&msg);
+    ::DispatchMessage(&msg);
+  }
 
   ::CoUninitialize();
   return EXIT_SUCCESS;
